refactor(templatetags): mover tags para chamados com compat layer em execucao
