+++ web/execucao/templatetags/execucao_ui.py
@@ -1,43 +1,11 @@
-# web/execucao/templatetags/execucao_ui.py
-from __future__ import annotations
+"""
+Compat layer para templatetags.
 
-from django import template
+Mantém imports legados como:
+- execucao.templatetags.execucao_ui
+- web.execucao.templatetags.execucao_ui
 
-register = template.Library()
+Source of truth: chamados.templatetags.execucao_ui
+"""
 
-COLOR_TO_CLASS = {
-    "SLATE": "bg-slate-200",
-    "BLUE": "bg-blue-500",
-    "EMERALD": "bg-emerald-500",
-    "VIOLET": "bg-violet-500",
-    "AMBER": "bg-amber-500",
-    "ROSE": "bg-rose-500",
-    "CYAN": "bg-cyan-500",
-    "LIME": "bg-lime-500",
-}
-
-
-@register.simple_tag
-def projeto_color_bar(projeto) -> str:
-    """
-    Retorna a classe tailwind da faixa de cor do projeto.
-    Fallback sempre para SLATE.
-    """
-    slug = (getattr(projeto, "cor_slug", None) or "SLATE").strip().upper()
-    return COLOR_TO_CLASS.get(slug, COLOR_TO_CLASS["SLATE"])
-
-
-@register.filter
-def get_item(mapping, key):
-    """
-    Acessa mapping.get(key) de forma segura no template.
-    Retorna None se 'mapping' não for um dict-like (ex.: string).
-    """
-    if mapping is None:
-        return None
-
-    getter = getattr(mapping, "get", None)
-    if getter is None:
-        return None
-
-    return getter(key)
+from chamados.templatetags.execucao_ui import *  # noqa: F403,F401
